FEM

PYMethod.FEPileModel

PYMethod.clear_boundary_conditions!

PYMethod.mass_matrix

PYMethod.solve!

PYMethod.stiffness_matrix

PYMethod.FEPileModel — Method

FEPileModel(bottom::Real, top::Real, nelements::Int) -> pile

Construct an object of the finite element model to simulate lateral behavior of pile. The ith Beam element can be accessed by pile[i]. The following values are vectors storing each nodal value.

Parameters

pile.E: Young's modulus of pile

pile.I: Second moment of area of pile

pile.D: Diameter of pile

Boundary conditions

pile.u: Lateral displacement

pile.θ: Angle of deflection (where θ can be typed by \theta<tab>)

pile.Fext: External lateral force

pile.Mext: External moment

The above vectors will be updated after solve! the problem.

Nodal variables

pile.coordinates: Coordinates of nodes

pile.F: Internal lateral force

pile.M: Internal moment

The internal lateral force and moment vectors will be updated after solve! the problem.

PYMethod.clear_boundary_conditions! — Method

clear_boundary_conditions!(::FEPileModel)

Clear boundary conditions. The parameters and p-y curves are remained.

PYMethod.mass_matrix — Method

mass_matrix(::Beam)

Construct mass matrix from Beam.

PYMethod.solve! — Method

solve!(::FEPileModel)

Solve the finite element problem.

PYMethod.stiffness_matrix — Method

stiffness_matrix(::Beam)

Construct element stiffness matrix from Beam.